+++ crates/notan_glow/src/texture.rs
@@ -3,7 +3,7 @@ use crate::{create_texture_from_html, GlowBackend};
 use glow::*;
 use notan_graphics::prelude::*;
 
-pub type TextureKey = <glow::Context as glow::HasContext>::Texture;
+pub(crate) type TextureKey = <glow::Context as glow::HasContext>::Texture;
 
 pub(crate) struct InnerTexture {
     pub texture: TextureKey,
@@ -143,6 +143,7 @@ pub(crate) unsafe fn create_texture(
     Ok(texture)
 }
 
+#[cfg(target_arch = "wasm32")]
 pub(crate) unsafe fn create_texture_from_html_image(
     gl: &Context,
     image: &web_sys::HtmlImageElement,
@@ -236,119 +237,4 @@ pub(crate) fn texture_internal_format(tf: &TextureFormat) -> u32 {
         TextureFormat::SRgba8 => glow::SRGB8_ALPHA8,
         _ => texture_format(tf),
     }
-}
-
-/// An image file to be uploaded to the gpu
-#[derive(Clone, Debug)]
-pub struct TextureSourceImage(pub Vec<u8>);
-
-impl TextureSource for TextureSourceImage {
-    fn upload(
-        &self,
-        device: &mut dyn DeviceBackend,
-        mut info: TextureInfo,
-    ) -> Result<(u64, TextureInfo), String> {
-        let backend: &mut GlowBackend = device
-            .as_any_mut()
-            .downcast_mut()
-            .ok_or("Invalid backend type".to_string())?;
-
-        let data = image::load_from_memory(&self.0)
-            .map_err(|e| e.to_string())?
-            .to_rgba8();
-
-        let pixels = if info.premultiplied_alpha {
-            premultiplied_alpha(&data)
-        } else {
-            data.to_vec()
-        };
-
-        info.bytes = Some(pixels);
-        info.format = TextureFormat::Rgba32;
-        info.width = data.width() as _;
-        info.height = data.height() as _;
-
-        let tex = unsafe { create_texture(&backend.gl, &info)? };
-        let id = backend.add_inner_texture(tex, &info)?;
-        Ok((id, info))
-    }
-}
-
-/// An image file to be uploaded to the gpu
-#[derive(Clone, Debug)]
-pub struct TextureSourceBytes(pub Vec<u8>);
-
-impl TextureSource for TextureSourceBytes {
-    fn upload(
-        &self,
-        device: &mut dyn DeviceBackend,
-        mut info: TextureInfo,
-    ) -> Result<(u64, TextureInfo), String> {
-        #[cfg(debug_assertions)]
-        {
-            let size = info.width * info.height * 4;
-            debug_assert_eq!(
-                self.0.len(),
-                size as usize,
-                "Texture bytes of len {} when it should be {} (width: {} * height: {} * bytes: {})",
-                self.0.len(),
-                size,
-                info.width,
-                info.height,
-                4
-            );
-        }
-
-        let backend: &mut GlowBackend = device
-            .as_any_mut()
-            .downcast_mut()
-            .ok_or("Invalid backend type".to_string())?;
-
-        let pixels = if info.premultiplied_alpha {
-            premultiplied_alpha(&self.0)
-        } else {
-            self.0.clone()
-        };
-
-        info.bytes = Some(pixels);
-
-        let tex = unsafe { create_texture(&backend.gl, &info)? };
-        let id = backend.add_inner_texture(tex, &info)?;
-        Ok((id, info))
-    }
-}
-
-/// An image file to be uploaded to the gpu
-#[derive(Clone, Debug)]
-pub struct TextureSourceHtmlImage(pub web_sys::HtmlImageElement);
-
-impl TextureSource for TextureSourceHtmlImage {
-    fn upload(
-        &self,
-        device: &mut dyn DeviceBackend,
-        mut info: TextureInfo,
-    ) -> Result<(u64, TextureInfo), String> {
-        let backend: &mut GlowBackend = device
-            .as_any_mut()
-            .downcast_mut()
-            .ok_or("Invalid backend type".to_string())?;
-
-        info.width = self.0.width() as _;
-        info.height = self.0.height() as _;
-
-        let tex = unsafe { create_texture_from_html_image(&backend.gl, &self.0, &info)? };
-        let id = backend.add_inner_texture(tex, &info)?;
-        Ok((id, info))
-    }
-}
-
-fn premultiplied_alpha(pixels: &[u8]) -> Vec<u8> {
-    pixels
-        .chunks(4)
-        .flat_map(|c| {
-            Color::from_bytes(c[0], c[1], c[2], c[3])
-                .to_premultiplied_alpha()
-                .rgba_u8()
-        })
-        .collect()
 }
