+++ crates/notan_web/src/lib.rs
@@ -12,5 +12,4 @@ mod files;
 mod audio;
 
 pub use backend::*;
-
-pub use notan_glow::{create_texture_from_html, TextureSourceHtmlImage, TextureSourceImage};
+pub use notan_glow::prelude::*;
