+++ crates/notan_winit/src/lib.rs
@@ -5,4 +5,4 @@ mod touch;
 mod window;
 
 pub use backend::*;
-pub use notan_glow::TextureSourceImage;
+pub use notan_glow::prelude::*;
